Fix up samples, move auth pieces into auth namespace
