+++ src/idunno.AtProto/Authentication/Models/BaseSessionResponse.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System.Text.Json.Serialization;
-using idunno.AtProto.Server;
 
 namespace idunno.AtProto.Authentication.Models
 {
