+++ test/idunno.AtProto.Serialization.Test/SessionDeserializationTests.cs
@@ -4,7 +4,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json;
 
-using idunno.AtProto.Models;
+using idunno.AtProto.Authentication.Models;
 
 namespace idunno.AtProto.Serialization.Test
 {
